+++ crates/tith-crypto/src/public_key.rs
@@ -4,7 +4,7 @@ use std::array::TryFromSliceError;
 
 use crate::raw as hydro;
 
-pub const PUBLIC_KEY_BYTES: usize = hydro::hydro_sign_PUBLICKEYBYTES as usize;
+pub const PUBLIC_KEY_BYTES: usize = hydro::hydro_sign_PUBLICKEYBYTES;
 
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
 pub struct PublicKey([u8; PUBLIC_KEY_BYTES]);
